requirements.txt
telebot==0.0.5
python-dotenv==1.0.0
jinja2==3.1.2